[Add] Added object Emoji
