+++ app/src/main/java/com/example/sauexpert/general_report/GeneralReportScreen.kt
@@ -5,20 +5,14 @@ import androidx.compose.foundation.layout.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.runtime.*
-import androidx.compose.runtime.R
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.example.sauexpert.model.TextOfTabData
-import com.example.sauexpert.patient_card_screen.BeginInspectionScreen
-import com.example.sauexpert.patient_card_screen.IndicatorPatientCardScreen
 import com.example.sauexpert.patient_card_screen.TabViewForPatientCardScreen
 import com.example.sauexpert.ui.theme.Gray30
 import com.example.sauexpert.widgets.compose.Toolbars.ActionToolBarWithSubtitle
-import com.example.sauexpert.widgets.compose.buttons.MainButtonsInRow
+
 
 @Composable
 fun GeneralReportScreen() {
