+++ app/src/main/java/com/example/sauexpert/signup_patient/CutomKeyboardOTPScreen.kt
@@ -3,8 +3,6 @@ package com.example.sauexpert.signup_patient
 import android.widget.Toast
 import androidx.compose.foundation.*
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.GridCells
-import androidx.compose.foundation.lazy.LazyVerticalGrid
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
